Fixed usage of Madgwick filter
As I suspected, the Madgwick filter doesn't accept references (durrrrr), but rather outputs quaternions. I didn't recognize it at first because the function type is void. I was expecting it to return a struct with the new rotations, but MadwickAHRS uses four floats as a quaternion and stores data in there between function calls - a cheap way of having a MadgwickAHRS class for all sakes and purposes.