Add preliminary version of SimplePID, for manual PID work.
SimplePID is designed to be a replacement for WPILib's lack of a
portable (not in the OS-sense) way of doing PID calculations, as
all WPILib PID-related classes (notably PIDController and CANTalon)
either manage such things internally or ask for a PIDSource
and PIDOutput object instead. As this is insufficient for manual
PID tasks (as anticipated in the 2016 FRC competition), this class
is designed to provide a remedy.